Separate the C2PO compile function into its own file
